feat: starting to implement edition for courses and categories
This is part of the question n4 from the Alura's challenge
Changes in course entity, controller, service, repository and JSP.
Changes in category entity, and controller

BREAKING CHANGE: The `CourseMapper.toEntity` previously accepts an `UpdateCourseFormDTO` as parameter, but this DTO will be discarded further and because of this method signature change to accept `NewCourseFormDTO dto` and `Long id`. In the persistence layer, `CourseRepository.existsByCode` has been removed (no more use) and new query methods have been added. Similar happens with `CourseService.update` (removed).