feat(registration): implementing logic in the controller to responde request
This is part of the question n6 from the Alura's challenge

BREAKING CHANGE: the `RegistrationControl.report()` now return a generic non-type `ResponseEntity` because in this way the application can return structured formal error messages with `ErrorItemDTO`. And by the way, the default system erro in `ErrorItemDTO` now  change from `system` to `error`.